requirements.txt
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0  
google-auth-oauthlib>=1.2.0
python-dateutil>=2.9.0 
pytesseract>=0.3.10         
SpeechRecognition>=3.10.4      
pyaudio>=0.2.14        
torch>=2.3.0  
nltk>=3.8.1                    
scikit-learn>=1.4.0      
numpy>=1.26.0          
requests>=2.32.0          
psutil>=5.9.8 
pycaw>=20240210             
comtypes>=1.4.4             
langdetect>=1.0.9      
deep-translator>=1.11.4    
pyttsx3>=2.90                  
beautifulsoup4>=4.12.3
lxml>=5.2.0              
selenium>=4.20.0      
webdriver-manager>=4.0.0        
pyperclip>=1.8.2           
Pillow>=10.3.0